requirements.txt
# Core
pyyaml>=6.0.1
python-dotenv>=1.0.1
pydantic>=2.6
requests>=2.31

# LLM + Agents
langchain>=0.2
langgraph>=0.2
google-generativeai>=0.7.2

# RAG / Vector
faiss-cpu>=1.7.4
tiktoken>=0.7.0

# Data / Analytics
pandas>=2.1
numpy>=1.26

# UI
streamlit>=1.34

# Testing
pytest>=8.0
pytest-cov>=4.1
